Creates nested navigator and enables navigation to details screen from each upcoming fixture card
